funding detail building

+++ Contract/funding.sol
@@ -1,9 +1,11 @@
 pragma solidity >=0.4.24 <=0.5.6;
+
 library SafeMath {
-  function sub(uint256 a, uint256 b) internal pure returns (uint256) {
-    assert(b <= a);
-    return a - b;
-  }
+    function sub(uint256 a, uint256 b) internal pure returns (uint256) {
+        assert(b <= a);
+        return a - b;
+    }
+
     function add(uint256 a, uint256 b) internal pure returns (uint256) {
         uint256 c = a + b;
         require(c >= a, "SafeMath: addition overflow");
@@ -11,10 +13,11 @@ library SafeMath {
         return c;
     }
 }
+
 contract Funding {
-    using SafeMath for uint;
-    mapping (address => uint256) private _balances;
-      
+    using SafeMath for uint256;
+    mapping(address => uint256) private _balances;
+
     address public owner;
     event Transfer(address indexed from, address indexed to, uint256 value);
     struct funding {
@@ -38,25 +41,32 @@ contract Funding {
     constructor() public {
         owner = msg.sender;
     }
-    
-    function createFunding(address restaurant,
+
+    function createFunding(
+        address restaurant,
         string memory foodname,
         uint256 endTime,
         uint256 totalAmount,
         uint256 price
     ) public {
-        fundingList.push(funding(restaurant, foodname,endTime, 0, totalAmount, 0, price));
+        fundingList.push(
+            funding(restaurant, foodname, endTime, 0, totalAmount, 0, price)
+        );
     }
-    
+
     /// 음식점 인덱스 찾기.
     function findTarget(
         address a,
         string memory f,
         uint256 money
-    ) public  returns (bool) {
+    ) public returns (bool) {
         uint256 result = 0;
         for (uint256 i = 0; i < fundingList.length; i++) {
-            if (fundingList[i].restaurant == a && keccak256(abi.encodePacked((fundingList[i].foodname))) == keccak256(abi.encodePacked((f)))) {
+            if (
+                fundingList[i].restaurant == a &&
+                keccak256(abi.encodePacked((fundingList[i].foodname))) ==
+                keccak256(abi.encodePacked((f)))
+            ) {
                 result = i;
                 break;
             }
@@ -66,9 +76,8 @@ contract Funding {
         }
         return false;
     }
-    
-    function UpdateValue(uint256 idx, uint256 money) public returns (bool)
-    {
+
+    function UpdateValue(uint256 idx, uint256 money) public returns (bool) {
         //uint256 userbalance = getBalance();
         if (fundingList[idx].amount < fundingList[idx].totalAmount) {
             if (getBalance() >= fundingList[idx].price) {
@@ -89,9 +98,9 @@ contract Funding {
         return false;
     }
 
-    function deposit() public payable {  
-         require(msg.sender == owner);
-    }   
+    function deposit() public payable {
+        require(msg.sender == owner);
+    }
 
     function getBalance() public view returns (uint256) {
         return address(this).balance;
@@ -101,7 +110,7 @@ contract Funding {
     //   _transfer(owner, fundingList[idx].restaurant, fundingList[idx].totalAmount);
     //   return true;
     // }
-        
+
     // function transferFromContract(address recipient, uint256 amount) public returns (bool) {
     //     _transfer(owner, recipient, amount);
     //     return true;
@@ -116,7 +125,7 @@ contract Funding {
     //     emit Transfer(sender, recipient, amount);
     // }
 
-     function transfer(uint256 amount) public returns (bool) {
+    function transfer(uint256 amount) public returns (bool) {
         msg.sender.transfer(amount);
         return true;
     }
@@ -124,6 +133,4 @@ contract Funding {
     // function getBalanceUser() internal view returns (uint256) {
     //     return address(msg.sender).balance;
     // }
-
-    
 }
